scraper/requirements.txt
# Web scraping
selenium>=4.12.0
pillow>=10.0.0

# Manipulation de données
pandas>=2.1.0
numpy>=1.27.0

# Base de données PostgreSQL
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Machine learning / recommandation
scikit-learn>=1.3.0
joblib>=1.3.0
